requirements.txt
google-adk
fastapi
uvicorn
httpx
pydantic
openai
streamlit